#5 Containerized ASP.NET API and use it inside docker-compose.
Added Dockerfile that builds a docker image for the ASP.NET API.

Made changes to the Program.cs so that when the application runs in production mode (which is the case inside the docker container, the web server hosting the API will listen on port 80). Otherwise it listens to random ports.

Added the ASP.NET API as a docker container inside the docker-compose, where before only the MySQL database ran.

Added configuration to the MySql database to use in production, and here the adress is inside of the docker container.