#103 created a new healthz endpoint in the api, which returns 200 OK if the application has started and can read from the database.
Also added healthcheck to the api in the docker-compose, using curl that calls the new healthz endpoint for up to 3 minutes and 30 seconds. It tries every 10 seconds and a soon as it gets a OK, it continues to start up the client. After 3 minutes and 30 seconds have passed without a OK from healthz, the docker-compose stops trying and reports an error.

Added curl to the API docker image, which is used in the health check.